# internal/app/ws/handler/ws.go
package handler

import (
	"context"
	"fmt"
	"net/http"

	"github.com/flamego/flamego"
	"github.com/gorilla/websocket"
	wshub "github.com/zet-plane/live-auction-backend/internal/app/ws/hub"
	"github.com/zet-plane/live-auction-backend/pkg/snowflake"
)

var (
	hub      *wshub.Hub
	upgrader = websocket.Upgrader{
		ReadBufferSize:  1024,
		WriteBufferSize: 1024,
		CheckOrigin:     func(r *http.Request) bool { return true },
	}
)

func Init(h *wshub.Hub) {
	hub = h
}

// ServeWS upgrades an authenticated live room WebSocket connection.
//
// @Summary 连接直播间 WebSocket
// @Tags websocket
// @Param room_id path string true "直播间 ID"
// @Param ticket query string true "通过 /api/v1/ws-ticket 签发的 ticket"
// @Success 101 {string} string "Switching Protocols"
// @Failure 400 {string} string "missing ticket or room_id"
// @Failure 401 {string} string "invalid or expired ticket"
// @Router /ws/v1/rooms/{room_id} [get]
func ServeWS(c flamego.Context, w http.ResponseWriter, r *http.Request) {
	roomID := c.Param("room_id")
	ticket := r.URL.Query().Get("ticket")
	if ticket == "" || roomID == "" {
		http.Error(w, "missing ticket or room_id", http.StatusBadRequest)
		return
	}

	key := fmt.Sprintf("ws:ticket:%s", ticket)
	userID, err := redisClient.GetDel(context.Background(), key).Result()
	if err != nil {
		http.Error(w, "invalid or expired ticket", http.StatusUnauthorized)
		return
	}

	wsConn, err := upgrader.Upgrade(w, r, nil)
	if err != nil {
		return
	}

	conn := wshub.NewConn("conn_"+snowflake.MakeUUID(), userID, roomID, wsConn, hub)
	hub.Register(conn)
	go conn.StartReadLoop()
	go conn.StartWriteLoop()
}