# internal/swaggerdocs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: Live auction e-commerce backend API.
  title: Live Auction Backend API
  version: "1.0"
paths:
...
  /ws/v1/rooms/{room_id}:
    get:
      parameters:
      - description: 直播间 ID
        in: path
        name: room_id
        required: true
        type: string
      - description: 通过 /api/v1/ws-ticket 签发的 ticket
        in: query
        name: ticket
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "400":
          description: missing ticket or room_id
          schema:
            type: string
        "401":
          description: invalid or expired ticket
          schema:
            type: string
      summary: 连接直播间 WebSocket
      tags:
      - websocket
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"